minor fixes
